[Feat] 유저모드 API 연결

See merge request s10-final/S10P31E205!99

+++ front_android/lib/src/model/user_mode_room.dart
@@ -33,7 +33,7 @@ class UserModeRoom {
       roomId: json['roomId'] ?? 0,
       name: json['name'] ?? '',
       capacity: json['capacity'] ?? 0,
-      distance: json['distance'] ?? 0,
+      distance: (json['distance'] as int).toDouble() ?? 0,
       status: json['status'] ?? UserModeRoomStatus.IN_PROGRESS,
       headcount: json['headcount'] ?? 1,
     );
